+++ lib/cocoapods/specification.rb
@@ -408,16 +408,8 @@ module Pod
       !source.nil? && !source[:local].nil?
     end
 
-    def local_path
-      Pathname.new(File.expand_path(source[:local]))
-    end
-
     def pod_destroot
-      if local?
-        local_path
-      else
-        config.project_pods_root + top_level_parent.name
-      end
+      config.project_pods_root + top_level_parent.name
     end
 
     def self.pattern_list(patterns)